refactor: extract axum handlers and state from main.rs into modules
